refactor(sumi): split the sumi controller into flows, jobs and locks
control.py held flow discovery, job handling and both protocol
controllers in one 1862 line module. The V2 payment work added to it,
so split it along its responsibilities before it grows further.

flows.py parses expose meta into MIP-002 flow items and checks
availability. jobs.py submits a job, heals a missing agent identifier
and reads job status back out of the event log. locks.py serves the
MIP-003 endpoints of a lock. control.py keeps SumiControl and
re-exports the moved helpers, so importers stay unchanged.

Code moved verbatim. _check_availability now lives in flows, so the
tests that patch HTTPXClient for it patch the flows module instead.
